backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==5.21.0
google-api-python-client==2.181.0
google-auth==2.40.3
gunicorn==23.0.0
psycopg[binary]==3.3.2
